refactor: :recycle: rename `useCreateForm` to `useCollectionForm` and obsoleted `CreateForm`
