+++ packages/react/src/react/views/collections/create/hooks/index.ts
@@ -1,2 +1,2 @@
+export { useCollectionForm } from '../../hooks/use-collection-form'
 export { useCollectionCreateMutation } from './use-collection-create-mutation'
-export { useCreateForm } from './use-create-form-with-default-values'
