+++ packages/react/src/react/views/collections/create/index.ts
@@ -1,2 +1,2 @@
 export { useCollectionCreate } from './context'
-export { useCollectionCreateMutation, useCreateForm } from './hooks'
+export { useCollectionCreateMutation, useCollectionForm } from './hooks'
