+++ packages/react/src/react/views/collections/create/view.tsx
@@ -3,15 +3,16 @@ import { type SubmitHandler } from 'react-hook-form'
 import { toast } from 'sonner'
 
 import { useCollectionCreate } from './context'
+import { useCollectionForm } from './hooks'
 import { useCollectionCreateMutation } from './hooks/use-collection-create-mutation'
 
-import { SubmitButton } from '../../../components'
+import { Form, SubmitButton } from '../../../components'
 import { useNavigation } from '../../../providers'
 
 export function CreateView() {
   const { navigate } = useNavigation()
   const {
-    components: { CreateForm, CreateFields },
+    components: { CreateFields },
   } = useCollectionCreate()
 
   const mutation = useCollectionCreateMutation({
@@ -24,12 +25,20 @@ export function CreateView() {
     },
   })
 
+  const form = useCollectionForm()
+
   const onSubmit: SubmitHandler<any> = (data: any) => mutation.mutateAsync(data)
 
   return (
-    <CreateForm onSubmit={onSubmit}>
-      <CreateFields disabled={mutation.isPending} />
-      <SubmitButton pending={mutation.isPending}>Create</SubmitButton>
-    </CreateForm>
+    <Form {...form}>
+      <form
+        noValidate
+        onSubmit={form.handleSubmit(onSubmit)}
+        className="flex flex-col gap-y-8 mt-16"
+      >
+        <CreateFields disabled={mutation.isPending} />
+        <SubmitButton pending={mutation.isPending}>Create</SubmitButton>
+      </form>
+    </Form>
   )
 }
